requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
packaging==24.2
pymongo==4.11
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
telebot==0.0.5
urllib3==2.3.0
Werkzeug==3.1.3
